// app/src/generated/models/Gk_activitesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Gk_activitesgk_action = {
  100000030: 'Demande recue',
  100000031: 'Changement de statut',
  100000032: 'Commentaire ajoute',
  100000033: 'Prorogation demandee',
  100000034: 'Prorogation approuvee',
  100000035: 'Fermeture'
} as const;
export type Gk_activitesgk_action = keyof typeof Gk_activitesgk_action;
export const Gk_activitesstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Gk_activitesstatecode = keyof typeof Gk_activitesstatecode;
export const Gk_activitesstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Gk_activitesstatuscode = keyof typeof Gk_activitesstatuscode;

export interface Gk_activitesBase {
  gk_action?: Gk_activitesgk_action;
  gk_activiteid: string;
  "gk_AuteurId@odata.bind"?: string;
  gk_commentaire?: string;
  gk_date: string;
  "gk_DemandeId@odata.bind"?: string;
  gk_name?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Gk_activitesstatecode;
  statuscode?: Gk_activitesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Gk_activites extends Gk_activitesBase {
  createdon?: string;
  gk_actionname?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  gk_auteurid?: object;
  _gk_auteurid_value?: string;
  gk_demandeid?: object;
  _gk_demandeid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}